+++ packages/fiori-annotation-api/src/utils/index.ts
@@ -1,0 +1,6 @@
+export { pathFromUri } from './path';
+export { PRIMITIVE_TYPE_NAMES } from './constants';
+export { getGenericNodeFromPointer } from './pointer';
+export { increaseIndent } from './indent';
+export { annotationReferenceToString } from './reference';
+export { compareByRange } from './range';