+++ packages/project-access/src/file/index.ts
@@ -1,2 +1,2 @@
-export { fileExists, readFile, readJSON, writeFile } from './file-access';
+export { fileExists, readFile, readJSON, updatePackageJSON, updateManifestJSON, writeFile } from './file-access';
 export { findBy, findFiles, findFilesByExtension, findFileUp, getFilePaths } from './file-search';